third_party/WebKit/LayoutTests/http/tests/inspector/websocket/websocket-frame-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
Connected.
Tests that WebSocketFrames are being sent and received by Web Inspector.

1- sent: test
1- sent: test
2- sent: exit
1- sent: test
2- sent: exit
3- received: test
1- sent: test
2- sent: exit
3- received: test
4- received: exit
